`rstan`: `io`: Fix type conversion warning
```c++
Found the following significant warnings:
  ../inst/include/rstan/io/r_ostream.hpp:48:20: warning: field precision specifier ‘.*’ expects argument of type ‘int’, but argument 2 has type ‘std::streamsize’ {aka ‘long int’} [-Wformat=]
  ../inst/include/rstan/io/r_ostream.hpp:71:21: warning: field precision specifier ‘.*’ expects argument of type ‘int’, but argument 2 has type ‘std::streamsize’ {aka ‘long int’} [-Wformat=]
```

+++ rstan/rstan/inst/include/rstan/io/r_ostream.hpp
@@ -45,7 +45,7 @@ namespace rstan {
       }
 
       virtual std::streamsize xsputn(const char_type* s, std::streamsize n) {
-        Rprintf("%.*s", n, s);
+        Rprintf("%.*s", static_cast<int>(n), s);
         return n;
       }
     };
@@ -68,7 +68,7 @@ namespace rstan {
       }
 
       virtual std::streamsize xsputn(const char_type* s, std::streamsize n) {
-        REprintf("%.*s", n, s);
+        REprintf("%.*s", static_cast<int>(n), s);
         return n;
       }
     };
